Changes to tfree#.g, tpre#.g, and tpost#.g
Deselecting a tool in FFF mode appears to remove the tool offset and can cause crashes if the tool is near Z=0 or immediately above a part. I've created tfree#.g files for all tools that check the mode, disable the spindles if it's already in CNC mode, and switch it to CNC mode if it's not.

Split tool changing functionality back into tpre#.g and tpost#.g files. When combined into a single file, the final return to stored position moves appeared to be causing a move below the previous Z coordinate and then back up to the stored coordinate. Perhaps this is due to the tool offset not being applied until after the move? Regardless, crashes were ocurring if a tool change was performed with the tool near Z=0 or immediately above a part.