Add Tool Offsets in Tfree#.g
Deselecting a tool while in FFF mode removes the tool offset. Because the H-Series machine uses negative Z offsets for its tools, this can cause crashes near Z=0. The updates here to tfree#.g files cause the tool offsets to be double applied before the tool is unselected. This will cause a small jog in +Z before a equivalent jog in -Z as the offset is removed. To avoid the "ratcheting" behavior in +Z when a tool is deselcted in CNC mode, FFF mode is asserted prior to the doubled tool offset move.