Updated tprime#.g Files
Previously, if a tprime#.g file is called while a tool other than number # is active, there was a superfluous movement as the new tool was moved into the active position and then rotated to the cleaning station. To avoid this superfluous movement, we now temporarily set the logical turret position to the expected new tool position before calling a standard tool change. Because the turret now appears to already be in the appropriate location, the turret and z-hop  movements in tpre#.g are skipped. After the tool change, we revert the logical turret position to reflect the actual position in reference to the newly selected tool. This allows tprime-universal.g to perform the appropriate movements for the priming and cleaning process.